NKartalBot/requirements.txt
flask
python-dotenv
requests
pillow
apscheduler
google-generativeai
openai
huggingface_hub
instagrapi
gunicorn